tests: coverage for string tensors
